Always use cerbero user inside docker
Instead of setting the username and uid on docker image creation, use a
script that executes as the container entry point. It creates a user
and group named cerbero, with the uid and gid of the user that runs
"docker run".

Also, reduce the number of steps needed to create the docker images.